requirements.txt
flask
openpyxl
pandas
faiss-cpu
sentence-transformers
llama-cpp-python
python-dotenv
numpy